+++ R/LandR-package.R
@@ -24,10 +24,10 @@
 #' @importFrom fpCompare %==% %>>% %<<%
 #' @importFrom ggplot2 aes coord_equal coord_sf element_blank element_text facet_wrap
 #' @importFrom ggplot2 geom_bar geom_hline geom_line geom_point labs geom_raster geom_ribbon geom_sf
-#' @importFrom ggplot2 ggplot guide_legend guides labs
+#' @importFrom ggplot2 ggplot ggtitle guide_legend guides labs
 #' @importFrom ggplot2 scale_color_distiller
 #' @importFrom ggplot2 scale_fill_distiller scale_fill_manual scale_fill_viridis_c scale_fill_viridis_d
-#' @importFrom ggplot2 scale_x_discrete
+#' @importFrom ggplot2 scale_linetype_manual scale_x_discrete
 #' @importFrom ggplot2 stat stat_summary sym theme theme_classic unit
 #' @importFrom ggpubr theme_pubr
 #' @importFrom ggspatial annotation_north_arrow layer_spatial north_arrow_minimal
@@ -52,13 +52,13 @@
 #' @importFrom SpaDES.tools spread2
 #' @importFrom stats as.formula complete.cases fitted glm na.omit predict quantile runif terms update vcov
 #' @importFrom terra app as.int cellFromRowCol cellFromXY classify coltab<- compareGeom
-#' @importFrom terra crop crs crs<- ext extract focalMat
-#' @importFrom terra intersect is.factor is.int is.points levels mask minmax NAflag<- ncell nlyr
-#' @importFrom terra project rast rasterize res rowColFromCell terraOptions values vect
+#' @importFrom terra crop crs crs<- ext extract focalMat intersect is.factor is.int is.points
+#' @importFrom terra levels mask minmax NAflag<- ncell nlyr project
+#' @importFrom terra rast rasterize res rowColFromCell terraOptions values vect
 #' @importFrom terra writeRaster xmax xmin ymax ymin xyFromCell
 #' @importFrom tidyterra geom_spatraster
 #' @importFrom tools file_path_sans_ext
-#' @importFrom utils capture.output combn count.fields data getFromNamespace install.packages
+#' @importFrom utils capture.output combn count.fields data getFromNamespace head install.packages
 #' @importFrom utils str tail untar
 ## usethis namespace: end
 NULL